requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
scrapy>=2.6.0
celery>=5.2.0
redis>=5.0.0
psycopg2-binary>=2.9.0
pymongo>=4.6.0
flask>=2.2.0
python-dotenv>=0.20.0

GitPython>=3.1.0
PyGithub>=1.58.0
pathlib>=1.0.1
tempfile>=1.0.0
shutil>=1.0.0

torch>=1.13.0
transformers>=4.21.0
datasets>=2.5.0
accelerate>=0.12.0
wandb>=0.13.0
scikit-learn>=1.1.0
pandas>=1.5.0
numpy>=1.23.0
openai>=0.27.0
fastapi>=0.85.0
uvicorn>=0.18.0
sqlalchemy>=1.4.0
pydantic>=1.10.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
python-multipart>=0.0.5

selenium>=4.8.0
playwright>=1.30.0
requests-html>=0.10.0

ast>=3.9.0
black>=22.0.0
flake8>=5.0.0
mypy>=1.0.0
autopep8>=2.0.4

langchain>=0.1.0
tiktoken>=0.3.0
chromadb>=0.3.0
sentence-transformers>=2.2.0

dataclasses>=0.6.0
typing-extensions>=4.4.0
aiohttp>=3.8.0
asyncio>=3.4.0
concurrent.futures>=3.1.0
multiprocessing>=0.70.0
logging>=0.4.0
json5>=0.9.0
yaml>=6.0.0
toml>=0.10.0
configparser>=5.3.0
urllib3>=1.26.0
certifi>=2022.12.0
charset-normalizer>=3.0.0
idna>=3.4.0

# Additional streaming and media processing
opencv-python>=4.8.0
pillow>=10.0.0
moviepy>=1.0.3
pydub>=0.25.0

# Enhanced security and network tools
scapy>=2.5.0
nmap>=0.0.1
python-nmap>=0.7.1
netaddr>=0.8.0
dnspython>=2.4.0

# Cloud and deployment automation
boto3>=1.29.0
azure-identity>=1.15.0
azure-mgmt-resource>=23.0.0
google-cloud-core>=2.4.0
kubernetes>=28.1.0
docker>=6.1.0

# Enhanced AI and ML capabilities
anthropic>=0.7.0
google-generativeai>=0.3.0
cohere>=4.37.0
huggingface-hub>=0.19.0

# Database connectors
sqlite3>=3.40.0

# API integrations
stripe>=7.8.0
twilio>=8.11.0
sendgrid>=6.10.0
mailgun>=1.0.0
paypalrestsdk>=1.13.3
razorpay>=1.3.0
cashfree-sdk>=4.0.0

# Development tools
pytest>=7.4.0
pytest-asyncio>=0.21.0
coverage>=7.3.0
pre-commit>=3.5.0

# Comprehensive IDE and platform support
streamlit>=1.28.0
plotly>=5.17.0
altair>=5.1.0
bokeh>=3.3.0

# Drag-and-drop builder engine
fabric>=5.3.0
paramiko>=3.3.0
invoke>=2.2.0

# Enhanced web scraping and automation
pyautogui>=0.9.54
keyboard>=0.13.5
mouse>=0.7.1

# Mobile and VTU system support
pysms>=1.0.0
africastalking>=1.2.5
flutterwave>=3.0.0

# UPI and Indian payment systems
paytmchecksum>=1.7.0
phonepe-sdk>=1.0.0

# Website builder and CMS integration
wordpress-api>=1.2.1
wix-api>=1.0.0
webflow-api>=1.0.0

# Design and graphics processing
cairosvg>=2.7.0
wand>=0.6.11
reportlab>=4.0.0
matplotlib>=3.8.0
seaborn>=0.12.0

# Advanced networking and security
cryptography>=41.0.0
pyopenssl>=23.3.0
pycryptodome>=3.19.0

# Real-time collaboration
socketio>=5.9.0
python-socketio>=5.9.0
eventlet>=0.33.0

# Enhanced data processing
openpyxl>=3.1.0
xlsxwriter>=3.1.0
python-docx>=1.1.0
pypdf2>=3.0.0

# Code generation and analysis
ast-decompiler>=1.0.0
astunparse>=1.6.3
isort>=5.12.0

# Template engines
jinja2>=3.1.0
mako>=1.3.0
chameleon>=4.5.0

# Enhanced API development
graphene>=3.3.0
strawberry-graphql>=0.213.0
ariadne>=0.21.0

# Monitoring and analytics
prometheus-client>=0.18.0
statsd>=4.0.1
newrelic>=9.2.0

# Enhanced deployment automation
ansible>=8.5.0
terraform-python>=0.1.0
pulumi>=3.90.0

# Advanced caching
memcached>=1.59.0
pymemcache>=4.0.0
redis-py-cluster>=2.1.3

# Enhanced testing
locust>=2.17.0
hypothesis>=6.88.0
factory-boy>=3.3.0

# Code quality and documentation
sphinx>=7.2.0
mkdocs>=1.5.0
pdoc>=14.1.0

# Enhanced file processing
python-magic>=0.4.27
filetype>=1.2.0
pathvalidate>=3.2.0

# Advanced scheduling
apscheduler>=3.10.0
celery-beat>=2.4.0
rq>=1.15.0

# Enhanced logging and monitoring
loguru>=0.7.0
structlog>=23.2.0
sentry-sdk>=1.38.0

# WebRTC and real-time communication
aiortc>=1.6.0
websockets>=12.0.0
